Add UsersControllerTest
